Fixed some redundancy
