+++ app/src/main/java/com/example/pygmyhippo/organizer/EventDB.java
@@ -93,33 +93,29 @@ public class EventDB extends DBHandler {
     }
 
     /**
-     * Gets a long-live download uris for Picasso to get image from Firebase Storage.
-     * @author James
-     * @param url String rul with gs:// link to Firebase Storage image.
-     * @param listener DBOnCompleteListener to call when query completes.
+     * This method will set the currently existing event to the new one with updated values
+     * @author Kori
+     * @param event The event we want to update
+     * @param listener The listener that initiates when the data is done updating
      */
-    public void getImageDownloadUrl(String url, StorageOnCompleteListener<Uri> listener) {
-        Log.d("DB", String.format("Getting storage reference for image url %s", url));
-        try {
-            // Try to get an image
-            StorageReference ref = storage.getReferenceFromUrl(url);
-            ref.getDownloadUrl().addOnCompleteListener(task -> {
-                if (task.isSuccessful()) {
-                    // Add the result to an array list and notify the listener
-                    Log.d("DB", String.format("Found image at %s", url));
-                    Uri uri = task.getResult();
-                    ArrayList<Uri> results = new ArrayList<>();
-                    results.add(uri);
-                    listener.OnCompleteStorage(results, 2, DBOnCompleteFlags.SUCCESS.value);
-                } else {
-                    Log.d("DB", String.format("Image download failed for %s", url));
-                }
-            });
-        } catch (IllegalArgumentException e) {
-            // Notify the listener with an error flag raised
-            Log.d("DB", String.format("IllegalArgumentException for url %s", url));
-            listener.OnCompleteStorage(new ArrayList<>(), 2, DBOnCompleteFlags.ERROR.value);
-        }
+    public void updateEvent(Event event, DBOnCompleteListener<Event> listener) {
+        db.collection("Events")
+                .document(event.getEventID())
+                .set(event)
+                .addOnCompleteListener(new OnCompleteListener<Void>() {
+                    @Override
+                    public void onComplete(@NonNull Task<Void> task) {
+                        ArrayList<Event> events = new ArrayList<>();
+                        if (task.isSuccessful()) {
+                            events.add(event);
+                            Log.d("DB", String.format("Successfully updated event with ID (%s).", event.getEventID()));
+                            listener.OnCompleteDB(events, 2, DBOnCompleteFlags.SUCCESS.value);
+                        } else {
+                            Log.d("DB", String.format("Error: Could not update event with ID (%s).", event.getEventID()));
+                            listener.OnCompleteDB(events, 2, DBOnCompleteFlags.ERROR.value);
+                        }
+                    }
+                });
     }
 
     /**
@@ -150,30 +146,6 @@ public class EventDB extends DBHandler {
                     } else {
                         Log.d("DB", String.format("Could not get event with organiser ID (%s).", accountID));
                         listener.OnCompleteDB(new ArrayList<>(), 3, DBOnCompleteFlags.ERROR.value);
-                    }
-                });
-    }
-
-    /**
-     * This method will set the currently existing event to the new one with updated values
-     * @author Kori
-     * @param event The event we want to update
-     * @param listener The listener that initiates when the data is done updating
-     */
-    public void updateEvent(Event event, DBOnCompleteListener<Event> listener) {
-        db.collection("Events")
-                .document(event.getEventID())
-                .set(event)
-                .addOnCompleteListener(new OnCompleteListener<Void>() {
-                    @Override
-                    public void onComplete(@NonNull Task<Void> task) {
-                        if (task.isSuccessful()) {
-                            Log.d("DB", String.format("Successfully updated event with ID (%s).", event.getEventID()));
-                            listener.OnCompleteDB(new ArrayList<>(), 4, DBOnCompleteFlags.SUCCESS.value);
-                        } else {
-                            Log.d("DB", String.format("Error: Could not update event with ID (%s).", event.getEventID()));
-                            listener.OnCompleteDB(new ArrayList<>(), 4, DBOnCompleteFlags.ERROR.value);
-                        }
                     }
                 });
     }
